+++ lib/src/image_file_view/io_preview.dart
@@ -25,7 +25,8 @@ class ImageFileView extends StatelessWidget {
     return Container(
       clipBehavior: Clip.antiAlias,
       decoration: BoxDecoration(
-        color: backgroundColor ?? Theme.of(context).colorScheme.background,
+        color: backgroundColor ??
+            Theme.of(context).colorScheme.surfaceContainerHighest,
         borderRadius: borderRadius ?? BorderRadius.zero,
       ),
       child: Uri.tryParse(imageFile.path!)?.scheme.startsWith('http') == true