+++ lib/src/widgets/default_draggable_item_widget.dart
@@ -2,10 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:multi_image_picker_view/multi_image_picker_view.dart';
 import 'package:multi_image_picker_view/src/multi_image_picker_controller_wrapper.dart';
 
-typedef DescriptionFieldCallback = Function(ImageFile, String);
-
 class DefaultDraggableItemWidget extends StatelessWidget {
-  DefaultDraggableItemWidget({
+  const DefaultDraggableItemWidget({
     super.key,
     required this.imageFile,
     this.fit = BoxFit.cover,
@@ -15,25 +13,10 @@ class DefaultDraggableItemWidget extends StatelessWidget {
     this.showCloseButton = true,
     this.closeButtonAlignment = Alignment.topRight,
     this.closeButtonIcon,
-    this.closeButtonBoxDecoration = const BoxDecoration(
-      color: Color(0x55AAAAAA),
-      shape: BoxShape.circle,
-    ),
+    this.closeButtonBoxDecoration,
     this.closeButtonMargin = const EdgeInsets.all(4),
     this.closeButtonPadding = const EdgeInsets.all(3),
-    this.showDescriptionField = false,
-    this.descriptionFieldText = "",
-    this.descriptionFieldReadOnly = false,
-    this.descriptionFieldHint = "Description",
-    this.descriptionFieldPadding =
-        const EdgeInsets.symmetric(horizontal: 8, vertical: 2),
-    this.descriptionFieldCallback,
-  }) {
-    descriptionController
-      ..text = descriptionFieldText
-      ..addListener(() => descriptionFieldCallback?.call(
-          imageFile, descriptionController.text));
-  }
+  });
 
   final ImageFile imageFile;
   final BoxFit fit;
@@ -44,13 +27,6 @@ class DefaultDraggableItemWidget extends StatelessWidget {
   final BoxDecoration? closeButtonBoxDecoration;
   final EdgeInsetsGeometry closeButtonMargin;
   final EdgeInsetsGeometry closeButtonPadding;
-  final TextEditingController descriptionController = TextEditingController();
-  final bool showDescriptionField;
-  final String descriptionFieldText;
-  final bool descriptionFieldReadOnly;
-  final String descriptionFieldHint;
-  final EdgeInsetsGeometry descriptionFieldPadding;
-  final DescriptionFieldCallback? descriptionFieldCallback;
 
   @override
   Widget build(BuildContext context) {
@@ -78,29 +54,21 @@ class DefaultDraggableItemWidget extends StatelessWidget {
                       .removeImage(imageFile),
                   child: Container(
                       padding: closeButtonPadding,
-                      decoration: closeButtonBoxDecoration,
-                      child:
-                          closeButtonIcon ?? const Icon(Icons.close, size: 18)),
+                      decoration: closeButtonBoxDecoration ??
+                          BoxDecoration(
+                            color: Theme.of(context)
+                                .colorScheme
+                                .surfaceContainerHighest
+                                .withOpacity(0.5),
+                            shape: BoxShape.circle,
+                          ),
+                      child: closeButtonIcon ??
+                          Icon(Icons.close,
+                              size: 18,
+                              color: Theme.of(context)
+                                  .colorScheme
+                                  .onSecondaryContainer)),
                 ),
-              ),
-            ),
-          ),
-        if (showDescriptionField)
-          Align(
-            alignment: Alignment.bottomCenter,
-            child: Padding(
-              padding: descriptionFieldPadding,
-              child: TextField(
-                decoration: InputDecoration(
-                  hintText: descriptionFieldHint,
-                  filled: true,
-                  fillColor: Colors.white70,
-                  isDense: true,
-                ),
-                readOnly: descriptionFieldReadOnly,
-                controller: descriptionController,
-                maxLines: 1,
-                style: Theme.of(context).textTheme.labelMedium,
               ),
             ),
           ),